refactor: 우승자 출력 기능 구조 변경
Car클래스와 직접 상호작용 하는 메소드 CarManager클래스 내부로 이동
